Move tests to custom/nouns folder & add tokenId == auctionId check
